+++ runelite-client/src/main/java/net/runelite/client/plugins/grounditems/GroundItemsOverlay.java
@@ -253,7 +253,7 @@ public class GroundItemsOverlay extends Overlay
 				}
 				else if (displayMode != PriceDisplayMode.OFF)
 				{
-					final int price = displayMode == PriceDisplayMode.GE
+					final long price = displayMode == PriceDisplayMode.GE
 						? item.getGePrice()
 						: item.getHaPrice();
